+++ CHANGELOG.md
@@ -4,6 +4,8 @@ ADDED:
 
 CHANGED:
 
+- Made the `RuntimeContainer` much-much more efficient
+
 REMOVED:
 
 FIXED: